Working on pointers on the stack

+++ res/C/samples/nora_sandler/array_4.c
@@ -1,6 +1,6 @@
 int main() {
     int my_array[3] = {1, 2, 3};
-    //int *my_pointer = my_array;
-    long *my_pointer = my_array;
+    int *my_pointer = my_array;
+    //long *my_pointer = my_array;
     return *my_pointer;
 }
